Making compilers happy
git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@189 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ scripting.c
@@ -74,7 +74,7 @@ Boolean ReadScriptParamsOnRead(void)
 	DescriptorKeyIDArray array = { NULLID };
 	int32 flags;
 	OSErr stickyError;
-	long v;
+	int32 v;
 
 	if (DescriptorAvailable(NULL)){ /* playing back.  Do our thing. */
 		token = OpenReader(array);
